+++ run/kubejs/server_scripts/prismac_test.js
@@ -14,9 +14,4 @@ ServerEvents.recipes(event => {
         .outputStatesNormal(PrismaticColor.GREEN)
         .EUt(GTValues.VA[GTValues.ZPM])
         .duration(40)
-
-    event.recipes.gtceu.sculk_vat("sculk_vat_test")
-        .itemInputs("minecraft:apple")
-        .outputFluids("minecraft:lava 3000")
-        .duration(10)
 })