+++ run/kubejs/server_scripts/sculk_vat_test.js
@@ -1,0 +1,6 @@
+ServerEvents.recipes(event => {
+    event.recipes.gtceu.sculk_vat("sculk_vat_test")
+        .itemInputs("minecraft:apple")
+        .outputFluids("minecraft:lava 3000")
+        .duration(10)
+})