+++ changelog.md
@@ -1,1 +1,4 @@
-# c
+# Moni Labs 0.14.1
+
+* Fix Sculk Vat not using its recipe modifier correctly
+* Adjust Sculk Vat EMI info to make it clearer